back-end/services/tts-&-voice-cloning-api/requirements.txt
Flask
Flask-Cors
soundfile
python-dotenv
